requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
# For Claude API
anthropic==0.7.8
# For Whisper (when implemented)
# openai-whisper==20231117
# For OCR (when implemented)
# pytesseract==0.3.10
# Pillow==10.0.1
# For development
python-dotenv==1.0.0